Implement MediaCapabilitiesDatabase with LevelDB
The LevelDB implementation supports the production use case of storing
video decode statistics for recall to answer MediaCapabilities queries.

This CL implements the interface and tests. A later CL will instantiate
the DB as part of the VideoDecodePerfHistory service for actual usage.

Kudos to mlamouri@ for first implementing this in
https://chromium-review.googlesource.com/c/chromium/src/+/625962

TEST: added to media_unittests
Bug: 695264

+++ media/mojo/services/media_capabilities_database.cc
@@ -11,8 +11,8 @@ MediaCapabilitiesDatabase::Entry::Entry(VideoCodecProfile codec_profile,
                                         int frame_rate)
     : codec_profile_(codec_profile), size_(size), frame_rate_(frame_rate) {}
 
-MediaCapabilitiesDatabase::Info::Info(uint32_t frames_decoded,
-                                      uint32_t frames_dropped)
+MediaCapabilitiesDatabase::Info::Info(uint64_t frames_decoded,
+                                      uint64_t frames_dropped)
     : frames_decoded(frames_decoded), frames_dropped(frames_dropped) {}
 
 }  // namespace media
